r23044: - use uint32_t for handle id's - include the session vuid in the SMB2 128-Bit wire handles   as SMB2 oplock breaks doesn't include a TID or VUID in the header   we need to make sure the handle is unique for the whole TCP connection
metze

+++ source/smb_server/handle.c
@@ -27,7 +27,7 @@
 /****************************************************************************
 init the handle structures
 ****************************************************************************/
-NTSTATUS smbsrv_init_handles(struct smbsrv_tcon *tcon, uint64_t limit)
+NTSTATUS smbsrv_init_handles(struct smbsrv_tcon *tcon, uint32_t limit)
 {
 	/* 
 	 * the idr_* functions take 'int' as limit,
@@ -47,7 +47,7 @@ NTSTATUS smbsrv_init_handles(struct smbsrv_tcon *tcon, uint64_t limit)
 find a handle given a handle id
 ****************************************************************************/
 static struct smbsrv_handle *smbsrv_handle_find(struct smbsrv_handles_context *handles_ctx,
-						uint64_t hid, struct timeval request_time)
+						uint32_t hid, struct timeval request_time)
 {
 	void *p;
 	struct smbsrv_handle *handle;
@@ -77,7 +77,7 @@ struct smbsrv_handle *smbsrv_smb_handle_find(struct smbsrv_tcon *smb_tcon,
 }
 
 struct smbsrv_handle *smbsrv_smb2_handle_find(struct smbsrv_tcon *smb_tcon,
-					      uint64_t hid, struct timeval request_time)
+					      uint32_t hid, struct timeval request_time)
 {
 	return smbsrv_handle_find(&smb_tcon->handles, hid, request_time);
 }
